// frontend/src/api/backend/backendSchemas.ts
export type RoleDto = {
  /**
   * Unique identifier of the role
   *
   * @example 65def4b2de3deddff04f79a9
   */
  _id: string;
  /**
   * Name of the role
   *
   * @example Insurance Admin
   */
  name: string;
};
// ...
export type ModuleDto = {
  /**
   * Unique identifier for the module
   *
   * @example 65e3579de0ca70fe8f2e148e
   */
  _id: ObjectId;
  /**
   * Name of the module
   *
   * @example Contract
   */
  name: string;
  /**
   * Code representing the module
   *
   * @example contract
   */
  code: string;
  /**
   * Status of the module
   *
   * @example true
   */
  status: boolean;
  /**
   * List of permissions for the module
   *
   * @example view-forms-menu
   * @example view-my-requests-tab
   */
  permissions: string[];
};

export type UserDataDto = {
  /**
   * Username of the user
   *
   * @example
   */
  username: string;
  /**
   * Email address of the user
   *
   * @example bala@nrl.com.auff
   */
  email: string;
  /**
   * Roles assigned to the user
   */
  roles: RoleDto[];
  /**
   * Modules accessible to the user
   */
  modules: ModuleDto[];
};

export type UserMetaResponseDto = {
  /**
   * Data payload containing user information
   */
  data: UserDataDto;
};